+++ src/main/java/at/jwe/snyder/data/record/OverviewLevel.java
@@ -1,3 +1,3 @@
 package at.jwe.snyder.data.record;
-public record OverviewLevel(String name, String size, Long levelId) {
+public record OverviewLevel(String name, String size, int levelId) {
 }
