+++ src/main/java/at/jwe/snyder/data/record/PlayerData.java
@@ -2,5 +2,5 @@ package at.jwe.snyder.data.record;
 
 import at.jwe.snyder.data.record.input.InputData;
 
-public record PlayerData(Long pathLength, int[] clusterData, InputData inputData) {
+public record PlayerData(int pathLength, int[] clusterData, InputData inputData) {
 }
