+++ src/main/java/at/jwe/snyder/data/record/PlayerResult.java
@@ -2,5 +2,5 @@ package at.jwe.snyder.data.record;
 
 import at.jwe.snyder.data.record.highscore.HighscoreEntry;
 
-public record PlayerResult(HighscoreEntry entry, Long levelId, PlayerData playerData) {
+public record PlayerResult(HighscoreEntry entry, int levelId, PlayerData playerData) {
 }
