+++ src/main/java/at/jwe/snyder/data/record/input/InputData.java
@@ -1,4 +1,4 @@
 package at.jwe.snyder.data.record.input;
 
-public record InputData(Long countUp, Long timeUp, Long countLeft, Long timeLeft, Long countRight, Long timeRight, Long countSpace, Long timeSpace) {
+public record InputData(int countUp, int timeUp, int countLeft, int timeLeft, int countRight, int timeRight, int countSpace, int timeSpace) {
 }
