+++ src/main/java/at/jwe/snyder/data/record/level/LayerInfo.java
@@ -1,4 +1,4 @@
 package at.jwe.snyder.data.record.level;
 
-public record LayerInfo(String name, int[] data, Long height, double opacity, String type, boolean visible, Long width, int x, int y) {
+public record LayerInfo(String name, int[] data, int height, double opacity, String type, boolean visible, int width, int x, int y) {
 }
