+++ src/main/java/at/jwe/snyder/data/record/level/Level.java
@@ -2,5 +2,5 @@ package at.jwe.snyder.data.record.level;
 
 import java.util.List;
 
-public record Level(List<LayerInfo> layers, String orientation, int tilewidth, int tileheight, Long height, Long width, List<Tileset> tilesets) {
+public record Level(List<LayerInfo> layers, String orientation, int tilewidth, int tileheight, int height, int width, List<Tileset> tilesets) {
 }
